requirements.txt
<<<<<<< HEAD
python-telegram-bot==22.5
flask==3.1.2
requests==2.32.5
sqlite3
=======
python-telegram-bot==20.7
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
>>>>>>> 4d33efe (Initial Ludo Bot setup)
